crypto-pulse/requirements.txt
# Ingestion
requests>=2.31.0
websocket-client>=1.7.0
schedule>=1.2.1
python-dotenv>=1.0.0
confluent-kafka>=2.3.0
newsapi-python>=0.2.7
websockets>=12.0
feedparser>=6.0.10

# Processing (Spark & Bronze/Silver)
pyspark>=3.5.0
delta-spark>=3.1.0

# Analytics (dbt Gold Layer)
dbt-spark>=1.7.0
dbt-postgres>=1.7.0

# ML & Data Science
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
tensorflow>=2.15.0
transformers>=4.35.0
torch>=2.1.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0

# Testing
pytest>=7.4.0
